migrate to indy apm-error-logging-plugin
