+++ apm-agent-plugins/apm-error-logging-plugin/src/main/java/co/elastic/apm/agent/errorlogging/AbstractLoggerErrorCapturingInstrumentation.java
@@ -22,11 +22,11 @@
  * under the License.
  * #L%
  */
-package co.elastic.apm.agent.error.logging;
+package co.elastic.apm.agent.errorlogging;
 
 import co.elastic.apm.agent.bci.TracerAwareInstrumentation;
-import co.elastic.apm.agent.bci.VisibleForAdvice;
 import co.elastic.apm.agent.impl.error.ErrorCapture;
+import co.elastic.apm.agent.sdk.state.GlobalThreadLocal;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.NamedElement;
 import net.bytebuddy.description.method.MethodDescription;
@@ -42,49 +42,39 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
 public abstract class AbstractLoggerErrorCapturingInstrumentation extends TracerAwareInstrumentation {
 
-    @SuppressWarnings({"WeakerAccess"})
-    @VisibleForAdvice
-    public static final ThreadLocal<Boolean> nestedThreadLocal = new ThreadLocal<Boolean>() {
-        @Override
-        protected Boolean initialValue() {
-            return Boolean.FALSE;
-        }
-    };
+    public static final GlobalThreadLocal<Boolean> nestedThreadLocal = GlobalThreadLocal.get(AbstractLoggerErrorCapturingInstrumentation.class, "nested", Boolean.FALSE);
 
     @Override
     public Class<?> getAdviceClass() {
         return LoggingAdvice.class;
     }
 
-    @Override
-    public boolean indyPlugin() {
-        return false;
-    }
-
     public static class LoggingAdvice {
 
-        @Advice.OnMethodEnter(suppress = Throwable.class)
-        public static void logEnter(@Advice.Argument(1) Throwable exception,
-                                    @Advice.Local("nested") boolean nested,
-                                    @Advice.Origin Class<?> clazz,
-                                    @Advice.Local("error") @Nullable ErrorCapture error) {
-            nested = nestedThreadLocal.get();
+        @Nullable
+        @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
+        public static Object[] logEnter(@Advice.Argument(1) Throwable exception,
+                                        @Advice.Origin Class<?> clazz) {
+            Boolean nested = nestedThreadLocal.get();
             if (!nested) {
-                error = tracer.captureException(exception, tracer.getActive(), clazz.getClassLoader());
+                ErrorCapture error = tracer.captureException(exception, tracer.getActive(), clazz.getClassLoader());
                 if (error != null) {
                     error.activate();
                 }
                 nestedThreadLocal.set(Boolean.TRUE);
+                return new Object[]{nested, error};
             }
+            return null;
         }
 
-        @Advice.OnMethodExit(suppress = Throwable.class)
-        public static void logExit(@Advice.Local("nested") boolean nested,
-                                   @Advice.Local("error") @Nullable ErrorCapture error) {
+        @Advice.OnMethodExit(suppress = Throwable.class, inline = false)
+        public static void logExit(@Advice.Enter @Nullable Object[] enter) {
+            Boolean nested = enter != null ? (boolean) enter[0] : null;
+            ErrorCapture error = enter != null ? (ErrorCapture) enter[1] : null;
             if (error != null) {
                 error.deactivate().end();
             }
-            if (!nested) {
+            if (nested != null && !nested) {
                 nestedThreadLocal.set(Boolean.FALSE);
             }
         }
