+++ apm-agent-plugins/apm-error-logging-plugin/src/main/resources/META-INF/services/co.elastic.apm.agent.sdk.ElasticApmInstrumentation
@@ -1,2 +1,2 @@
-co.elastic.apm.agent.error.logging.Slf4jLoggerErrorCapturingInstrumentation
-co.elastic.apm.agent.error.logging.Log4j2LoggerErrorCapturingInstrumentation
+co.elastic.apm.agent.errorlogging.Slf4jLoggerErrorCapturingInstrumentation
+co.elastic.apm.agent.errorlogging.Log4j2LoggerErrorCapturingInstrumentation
