+++ apm-agent-plugins/apm-log-correlation-plugin/src/test/java/co/elastic/apm/agent/mdc/MdcActivationListenerIT.java
@@ -27,8 +27,8 @@ package co.elastic.apm.agent.mdc;
 import co.elastic.apm.agent.MockReporter;
 import co.elastic.apm.agent.bci.ElasticApmAgent;
 import co.elastic.apm.agent.configuration.SpyConfiguration;
-import co.elastic.apm.agent.error.logging.Log4j2LoggerErrorCapturingInstrumentation;
-import co.elastic.apm.agent.error.logging.Slf4jLoggerErrorCapturingInstrumentation;
+import co.elastic.apm.agent.errorlogging.Log4j2LoggerErrorCapturingInstrumentation;
+import co.elastic.apm.agent.errorlogging.Slf4jLoggerErrorCapturingInstrumentation;
 import co.elastic.apm.agent.impl.ElasticApmTracerBuilder;
 import co.elastic.apm.agent.impl.ElasticApmTracer;
 import co.elastic.apm.agent.impl.transaction.Transaction;
